Initial commit to refactor tags to support multiple resource types
