+++ amundsen_application/static/js/ducks/rootSaga.ts
@@ -45,7 +45,6 @@ import {
 
 // TableDetail
 import { updateTableOwnerWatcher } from './tableMetadata/owners/sagas';
-import { updateTableTagsWatcher } from './tableMetadata/tags/sagas';
 import {
   getTableDataWatcher,
   getColumnDescriptionWatcher,
@@ -57,7 +56,7 @@ import {
 } from './tableMetadata/sagas';
 
 // Tags
-import { getAllTagsWatcher } from './allTags/sagas';
+import { getAllTagsWatcher, updateTableTagsWatcher } from './tags/sagas';
 
 // User
 import { getLoggedInUserWatcher, getUserOwnWatcher, getUserReadWatcher, getUserWatcher } from "./user/sagas";
@@ -96,6 +95,7 @@ export default function* rootSaga() {
     getPopularTablesWatcher(),
     // Tags
     getAllTagsWatcher(),
+    updateTableTagsWatcher(),
     // TableDetail
     getTableDataWatcher(),
     getColumnDescriptionWatcher(),
@@ -105,7 +105,6 @@ export default function* rootSaga() {
     updateColumnDescriptionWatcher(),
     updateTableDescriptionWatcher(),
     updateTableOwnerWatcher(),
-    updateTableTagsWatcher(),
     // User
     getLoggedInUserWatcher(),
     getUserWatcher(),
