+++ amundsen_application/static/js/ducks/tableMetadata/api/v0.ts
@@ -22,29 +22,8 @@ import {
   createOwnerUpdatePayload, createOwnerNotificationData, shouldSendNotification
 } from './helpers';
 
-export function getTableTags(tableKey: string) {
-  const tableParams = getTableQueryParams(tableKey);
-  return axios.get(`${API_PATH}/table?${tableParams}`)
-  .then((response: AxiosResponse<TableDataAPI>) => {
-    return getTableTagsFromResponseData(response.data);
-  });
-}
 
-/* TODO: Typing this method generates redux-saga related type errors that needs more dedicated debugging */
-export function updateTableTags(tagArray, tableKey: string) {
-  const updatePayloads = tagArray.map((tagObject) => {
-    return {
-      method: tagObject.methodName,
-      url: `${API_PATH}/update_table_tags`,
-      data: {
-        key: tableKey,
-        tag: tagObject.tagName,
-      },
-    }
-  });
-  return updatePayloads.map(payload => { axios(payload) });
-}
-
+// TODO - Revisit the necessity of filtering out tags from main metadata
 export function getTableData(tableKey: string, index?: string, source?: string ) {
   const queryParams = getTableQueryParams(tableKey, index, source);
   return axios.get(`${API_PATH}/table?${queryParams}`)
